backend/requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.3
DateTime==5.5
distro==1.9.0
Django==5.2.1
django-allauth==65.8.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
etelemetry==0.3.1
filelock==3.18.0
FITX==1.0.1
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
looseversion==1.3.0
lxml==6.0.0
mysqlclient==2.2.7
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.2
oauthlib==3.2.2
openai==1.81.0
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
playwright==1.53.0
proto-plus==1.26.1
protobuf==6.31.0
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.1
pyee==13.0.0
PyJWT==2.9.0
PyMuPDF==1.26.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
scipy==1.16.1
setuptools==80.8.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
zope.interface==7.2
